ocr_processor/requirements.txt
kafka-python==2.0.2 
python-dotenv==1.0.1 
psycopg2-binary==2.9.9 
PyMuPDF==1.24.10 
pytesseract==0.3.10 
Pillow==10.4.0 
google-generativeai==0.8.3
uuid